[Decomp][Player] Cohesion split: extract chat methods to PlayerChat.cpp
